macros: split lib.rs into parse and expand modules
Mechanical split of the 822-line proc-macro lib.rs. The entry points
stay in the crate root (a proc-macro requirement) with their public
docs; parse.rs takes SubscriberArgs/WorkersArg and the syntactic
inspectors (source type recovery, Result/Vec return shapes, doc
comments); expand.rs takes the four #[subscriber] definition forms
around HandlerParts. The small #[app] and derive(Message) expansions
stay beside their entries in lib.rs.
